create constants for test
