+++ packages/ui-helpers/src/__tests__/Footer.spec.tsx
@@ -9,6 +9,12 @@ import {
 } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import { CLICK_FEEDBACK_FOOTER, MixpanelProvider, useTracker } from "tracker";
+import {
+  CONFIG,
+  EMPTY_TOKEN,
+  TOKEN,
+  eventTriggerByText,
+} from "./testConstants";
 
 describe("Testing for Footer", () => {
   beforeEach(() => {
@@ -16,11 +22,8 @@ describe("Testing for Footer", () => {
   });
 
   it("calls Mixpanel functions without token", () => {
-    const token = "";
-    const config = { debug: true, ip: false };
-
     const wrapper = ({ children }: { children: JSX.Element }) => (
-      <MixpanelProvider token={token} config={config}>
+      <MixpanelProvider token={EMPTY_TOKEN} config={CONFIG}>
         {children}
       </MixpanelProvider>
     );
@@ -42,16 +45,12 @@ describe("Testing for Footer", () => {
   });
 
   it("calls Mixpanel functions with token", () => {
-    const token = "testToken";
-    const config = { debug: true, ip: false };
     render(<Footer />);
     const expectedText = "Feedback";
-    const textElement = screen.getByText(expectedText);
-    expect(textElement).toBeInTheDocument();
-    fireEvent.click(textElement);
+    eventTriggerByText(expectedText);
 
     const wrapper = ({ children }: { children: JSX.Element }) => (
-      <MixpanelProvider token={token} config={config}>
+      <MixpanelProvider token={TOKEN} config={CONFIG}>
         {children}
       </MixpanelProvider>
     );
